Milestone 0.8 (original deadline 15th July)
Moved most stuff over from differentiation/SQD-MNP report and the paper.
References have been left blank (tagged with REF) to be put in later. Questions
and notes-to-self in red. This is still very (VERY) rough. Appendices still
need to be filled out and modified, though they are referenced to. Also still
need to copy the population inversion part over from the paper.